Consolidate cutout and standard photo post.
